feat: accessory form parity + accessories filter by type
Add 5 missing fields (serialNumber, acquisitionDate, initialRoundCount,
notes, imageUrl) to the inline Create New Accessory form in the build
configurator modal, matching the full /accessories/new page. All new
fields are included in the POST body to /api/accessories.

Extract accessories page into a Server Component + Client Component
pattern, adding a filter-by-type chip row above the accessories grid
so users can narrow the list by slot type (Optic, Barrel, etc.).
